logging: get rid of heavyweight 'feature-laden' logging facility
Removes module-scope prefixes, WARN prefixes, and timestamp prefixes.  Most of these things were hold-overs from things other people added to my library as I've been bringing it along with me from project to project... and while I _kind_ of liked the module scope prefix, it was a pain when it came to C++ classes, or printing multi-line text where prefixes interfere with other things.  Adding prefixes manually, as it turns out, is pretty damn easy too.  Easier, in most cases.  So it's gone!
