+++ CraftEngine/ajek-framework/h/x-ScopedMalloc.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include "x-stdlib.h"
+#include "x-stdalloc.h"
 #include "x-assertion.h"
 
 // --------------------------------------------------------------------------------------
