+++ CraftEngine/ajek-framework/h/x-stdlib.h
@@ -8,7 +8,6 @@
 
 #include <cstring>      // needed for memset
 #include <type_traits>
-#include <cstdio>
 
 #define serialization_assert(T) \
     static_assert(xIs_trivially_copyable(T), "Cannot serialize non-POD object.")
@@ -52,134 +51,6 @@ public:
     }
 };
 
-
-
-// ======================================================================================
-//  xCustomAllocator  (interface)
-// ======================================================================================
-class xCustomAllocator {
-public:
-    virtual void*       alloc       (size_t size)               =0;
-    virtual void*       realloc     (void* ptr, size_t size)    =0;
-    virtual void        free        (void* ptr)                 =0;
-};
-
-class xDefaultCustomAllocator : public xCustomAllocator {
-public:
-    void*       alloc       (size_t size)               override;
-    void*       realloc     (void* ptr, size_t size)    override;
-    void        free        (void* ptr)                 override;
-};
-
-extern xDefaultCustomAllocator  g_DefaultCustomAllocator;
-
-// ======================================================================================
-//  xBaseStream (abstract class)
-// ======================================================================================
-class xBaseStream
-{
-    NONCOPYABLE_OBJECT( xBaseStream );
-
-protected:
-    union {
-        // We cheat her a bit -- m_fd strictly speaking should be -1 for an empty descriptor.
-        // The API manually sets -1 to zero (0) within OpenFile() to allow !nullptr checks
-        // against m_fp to satisfy either underlying file system.  This is ok because posix
-        // fd==0 is reserved as stdin anyway, and will never be returned from open(). --jstine
-
-        FILE*       m_fp;               // non-NULL implies data is sourced from file
-        int         m_fd;
-    };
-
-    size_t              m_dataPos       = 0;
-    size_t              m_dataLength    = 0;
-    bool                m_isBuffered    = false;            // if TRUE, use m_fp else use m_fd
-    xCustomAllocator*   m_allocator     = &g_DefaultCustomAllocator;
-
-protected:
-    xBaseStream() {
-        m_fp            = nullptr;
-    }
-
-public:
-    virtual             ~xBaseStream    () throw()=0;
-    virtual void        Close           ();
-    virtual const void* GetBufferPtr    (size_t size) const=0;
-    virtual void*       GetBufferPtr    (size_t size)=0;
-            x_off_t     Seek            (x_off_t pos, int whence);
-            x_off_t     Tell            ()              const;
-
-            void        SetCustomAlloc  (xCustomAllocator& allocator)   { m_allocator = &allocator; }
-            bool        IsMemory        ()              const   { return  !m_fp;  }
-            bool        IsFile          ()              const   { return !!m_fp; }
-            bool        IsOK            ()              const   { return !!m_fp || !!m_dataLength; }
-            FILE*       GetFILE         ()                      { return m_isBuffered ? m_fp : nullptr; }
-};
-
-inline xBaseStream::~xBaseStream() throw() {}
-
-
-// ======================================================================================
-//  xStreamReader (class)
-// ======================================================================================
-class xStreamReader : public xBaseStream
-{
-    typedef xBaseStream __parent;
-
-protected:
-    const u8*   m_data;
-
-public:
-                        xStreamReader   ()          { }
-    virtual             ~xStreamReader  () throw() { Close(); }
-    virtual void        Close           ()                          override;
-    virtual const void* GetBufferPtr    (size_t size) const         override;
-    virtual void*       GetBufferPtr    (size_t size)               override;
-            bool        OpenFile        ( const xString& filename );
-            bool        OpenFileBuffered( const xString& filename );
-            bool        OpenMem         ( const void* mem, size_t buffer_length );
-            bool        Read            ( void* dest, ssize_t bytes );
-
-    // Valid for both file and data streams in xStreamReader.
-            size_t      GetLength       ()              const   { return m_dataLength; }
-
-    template< typename T >
-    bool Read( T& dest )
-    {
-        serialization_assert(T);
-        return Read( &dest, sizeof(T) );
-    }
-};
-
-struct xStatInfo
-{
-    u32         st_mode;
-    s64         st_size;
-
-    time_t      time_accessed;
-    time_t      time_modified;
-    time_t      time_created;
-
-    bool IsFile     () const;
-    bool IsDir      () const;
-    bool Exists     () const;
-
-};
-
-
-extern void     xMalloc_Check       ();
-extern void*    xMalloc             (size_t sz);
-extern void*    xCalloc             (size_t numItems, size_t sz);
-extern void*    xRealloc            (void* srcptr, size_t sz);
-extern void     xFree               (void *ptr);
-extern void*    xMalloc_Aligned     (size_t sz, u32 align);
-extern void*    xRealloc_Aligned    (void* srcptr, size_t sz, u32 align);
-extern void     xFree_Aligned       (void* ptr);
-extern void     xMalloc_Report      ();
-extern void     xMalloc_ReportDelta ();
-
-#define placement_new(T)        new (xMalloc(sizeof(T))) T
-
 template<typename T>
 void placement_delete(T* ptr)   { (ptr)->~T(); xFree(ptr); }
 
@@ -195,41 +66,9 @@ extern void         xMemCopyQwc_WrappedSrc  (u128* dest,        const u128* srcB
 extern void         xMemCopyShortQwc        (void* dest, const void* src, uint lenQwc);
 extern void         xMemCopyShortQwc_NT     (void* dest, const void* src, uint lenQwc);
 
-extern xStatInfo    xFileStat               (const xString& path);
-extern bool         xFileRename             (const xString& src, const xString& dst);
-extern bool         xCreateDirectory        (const xString& dir);
-extern FILE*        xFopen                  (const xString& fullpath, const char* mode);
-
-extern void         xFileSetSize            (int fd, size_t filesize);
-extern bool         xFgets                  (xString& dest, FILE* stream);
-
 extern bool         xEnvironExists          (const xString& varname);
 extern xString      xEnvironGet             (const xString& varname);
 extern void         xEnvironSet             (const xString& varname, const xString& value, bool overwrite=1);
-
-
-// Performs cleanup of existing object pointer (if non-null) and creates a new object in its place.
-// Memory is allocated if the provided pointer is null.
-template< typename T >
-__xi T* xMallocNew(T* &ptr)
-{
-    if (ptr) {
-        ptr->~T();
-    }
-    else {
-        ptr = (T*)xMalloc(sizeof(*ptr));
-    }
-    return new (ptr) T();
-}
-
-// Memory is allocated if the provided pointer is null.  if pointer is non-null, then no action is taken.
-template< typename T >
-__xi T* xMallocT(T* &ptr)
-{
-    static_assert(!std::has_virtual_destructor<T>::value, "Non-trivial type has meaningful destructor.  Use xMallocNew<T> instead.");
-    //static_assert(!std::is_trivially_copyable<T>::value, "Non-trivial type has meaningful destructor.  Use xMallocNew<T> instead.")
-    return ptr ? ptr : xMalloc(sizeof(*ptr));
-}
 
 template< typename T >
 inline void xMemMove(T* dest, size_t destLen, const T* src, size_t srcLen)
