+++ CraftEngine/src/test-spline.cpp
@@ -10,8 +10,6 @@
 
 #include "Bezier2d.inl"
 
-DECLARE_MODULE_NAME("spline");
-
 static const int const_zval = 0.1f;
 
 static const int        numStepsPerCurve        = 10;
